Menu is now is a new file for both x11 and wayland
