+++ src/os/unix/x11.rs
@@ -9,7 +9,7 @@ use x11_dl::xlib;
 
 use crate::error::Error;
 use crate::Result;
-use crate::{CursorStyle, MenuHandle, MenuItem, MenuItemHandle, UnixMenu, UnixMenuItem};
+use crate::{CursorStyle, MenuHandle, UnixMenu};
 
 use std::ffi::CString;
 use std::mem;
@@ -19,6 +19,8 @@ use std::ptr;
 
 use crate::buffer_helper;
 use crate::mouse_handler;
+
+use super::common::Menu;
 
 // NOTE: the x11-dl crate does not define Button6 or Button7
 const Button6: c_uint = xlib::Button5 + 1;
@@ -1103,58 +1105,3 @@ impl Drop for Window {
     }
 }
 
-pub struct Menu {
-    pub internal: UnixMenu,
-}
-
-impl Menu {
-    pub fn new(name: &str) -> Result<Menu> {
-        Ok(Menu {
-            internal: UnixMenu {
-                handle: MenuHandle(0),
-                item_counter: MenuItemHandle(0),
-                name: name.to_owned(),
-                items: Vec::new(),
-            },
-        })
-    }
-
-    pub fn add_sub_menu(&mut self, name: &str, sub_menu: &Menu) {
-        let handle = self.next_item_handle();
-        self.internal.items.push(UnixMenuItem {
-            label: name.to_owned(),
-            handle,
-            sub_menu: Some(Box::new(sub_menu.internal.clone())),
-            id: 0,
-            enabled: true,
-            key: Key::Unknown,
-            modifier: 0,
-        });
-    }
-
-    fn next_item_handle(&mut self) -> MenuItemHandle {
-        let handle = self.internal.item_counter;
-        self.internal.item_counter.0 += 1;
-        handle
-    }
-
-    pub fn add_menu_item(&mut self, item: &MenuItem) -> MenuItemHandle {
-        let item_handle = self.next_item_handle();
-        self.internal.items.push(UnixMenuItem {
-            sub_menu: None,
-            handle: self.internal.item_counter,
-            id: item.id,
-            label: item.label.clone(),
-            enabled: item.enabled,
-            key: item.key,
-            modifier: item.modifier,
-        });
-        item_handle
-    }
-
-    pub fn remove_item(&mut self, handle: &MenuItemHandle) {
-        self.internal
-            .items
-            .retain(|ref item| item.handle.0 != handle.0);
-    }
-}
